bugfix: price convertion on cart & checkout pages
+ currency icons in header

+++ frontend/web/js/checkout.js
@@ -29,7 +29,7 @@ $(function () {
 
         $('#shipping-price').html(amount);
 
-        $('.cart-total .val').html(parseInt(cartSubTotal) + parseInt(raw_amount));
+        $('.cart-total .val').html(parseFloat(cartSubTotal) + parseFloat(raw_amount));
     });
 
     $('.zones_select').change(function () {
